backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.23.2
sqlalchemy>=2.0.22
psycopg2-binary>=2.9.9
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
minio>=7.1.15
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
alembic>=1.12.0
email-validator>=2.0.0

# Microservice integration
requests>=2.31.0

# Logging and monitoring
python-json-logger>=2.0.7